requirements.txt
fpdf==1.7.2
gitdb==4.0.10
GitPython==3.1.31
Pygments==2.14.0
smmap==5.0.0
